Ye shalt not find a better string comparison routine
The input word is written to a word-buffer backwards, to match the heap layout.

The length of the word is calculated not by incrementing C for every byte, but by using the maximum number of characters allowed to reverse the bytes-remaining count!

`findSymbol` is a work of art.

Word lengths are compared and links followed with a minimum of instructions and without having to reload the word-length unless and actual string comparison is required.

CPD is used to compare and decrement during string comparison with the word-length, BC, used to snap back from an arbitrary point in the word (where mismatch occurs), back to the link-field!