Slight rework of instruction parsing
When a character match is made that results in an opcode, we need to check that the next character is whitespace/eof and error otherwise.

The hi/lo absolute address for branch jumps has been changed to an offset. Whilst slower, the table is no longer bound to being <$8000 in RAM.