Minor changes to symbol search
When parsing an expression, symbols are flattened into their values; a reference to the symbol isn't needed -- unless a forward-reference is used.
When this happens, what should we do? We can't append a new label definition to the heap in the middle of writing an expression to the heap.
Including the full label name would be inefficient.

Expression tokenising is going to be moved onto the bottom of the stack area so that we can create new label placeholders in the heap and just use an address to refer to them in deferred expressions.

I'm still not 100% certain that this is the right apporach -- if we build the expr forwards, we will have to copy it to the heap backwards. Where do we evaluate the expression? Some copying will be needed either way.

Also, 2024.