Improvements to `readWord`
A terminator is not needed on the word buffer, so it's been adjusted to run from $00 to $1F instead of $20 (where expr bytecode will now start)

The character reading loop no longer needs a counter as we can use the lo-byte underflowing as the stop indicator.
The end of the routine is simplified by not retutning `DE` set to the top of the word-buffer; we now do that in `findSymbol`.
`findSymbol` was improved by only needing to reset the lo-byte of  `DE` (the word-buffer addr) instead of a costly 20-cycles for a `ld DE, [word]`.