Add deferred flag to labels
Using a bit on the word length is too fiddly to preserve until after the full string comparison, so I just use a byte before the value. This will be redundant for constants, but perhaps it can be used for something else *shrug*

The deferred state is return in the carry flag after `findSymbol`, although this process might change -- a label that is deferred is still considered missing as far as an expression is concerned, but we need the addr for the deferred label token.

For efficiency there needs to be a way to update a deferred label's value when defining a new label. This could be done with a flag but it needs to be preserved until after the symbol match.