Scope parsing part n of many...
An expression at the start of a line, before any keyword or instruction, sets the program-counter.
This expression cannot contain any forward-references and in general we need to be able to return a flag for if an expression contains forward references.

`PushLabel` now stores the forward-reference flag in a way taht will return A=0, zf=1 AND cf=1 when `findSymbol` is used on a forward-reference.