Update a forward-reference when the label is defined
If a label that is not know yet is referred to in an expression, a placeholder symbol is put on the heap (for the deferred expression to point to), once that label is defined we go back and update the value to the program-counter.

`findSymbol` will return the address of the symbol's flag-byte+value if the symbol is deferred, avoiding some pointer arithmetic.