Swap direction of expr tokens, again
Expr bytecode will be written from the middle of the stack area (i.e. $FC7F) downwards.
Whilst this limits the stack size to a fixed 128 bytes (I hope this is big enough!), we can easily copy the expr tokens to the heap as the directions match.
Due to the word-buffer at $FC00..$FC1F, this gives about 96 bytes for the expr tokens which should be more than enough for an 80-col line.