Rejigging parser
The parser needs to be able to return to the body context after parsing an instruction or keyword or to restart at a new line.

Parsing instructions or keywords could consume the first letter of the next word and we need to be prepared to be able to re-use any printable charcter or to skip ahead for whitespace.