Undo some previous work
In instances where routines can't return the newline/end-of-file flag from `readChar`, like `getSymbol`, we need to check for end-of-file when trying to return to line/body context.
`readChar` will now guarantee 0 = end of file so that `[char]` can be used at any time to check for end of file.