Fix broken instruction parsing
Forgot that the sign-bit is the result of the comparison, and not copied from the source as it is on the 6502!
We weren't checking the hi-bit of characters in the opcode table correctly.
Also, caught two branches missing ends.
A bunch of instructions now assemble (although I don't know if the output is correct!)