Fix comment oversight
A comment when an parameter was expected would throw off the parser that was looking for a newline or end-of-file to mark the end of an expression.
Rather than change the parser to check for comments anytime it also checks for newline/EOF, the routine for skipping spaces now skips comments to the requisite newline/EOF meaning the parser never sees comments ever.
A parameter is left off of the test file to demonstrate that it now expects a parameter rather than beliving the expression is invalid.