Begin constant parsing
Label and constant behaviour is different -- labels cannot be redefined but they can be deferred (forward-reference) whereas constants can be redefined but the expression must be constant (no forward-references).
Here we read the constant name but don't search for it ahead-of-time; it doesn't matter if it already  exists or not, we need the value first before we can add/modify the symbol record.