Calc expr length when parsing
When an expr is tokenised, we now record the length in case the expr has to be deferred and we need to copy it to the heap.
The hi-bit of the length is set to indicate the expr contains deferred labels that cannot be resolved -- this might change to carry flag or something else later.