Inline tokenExpr
With a minor caveat, we can simplify things by having expr parsing automatically defer the expr to the heap if needed.
For constant defines and if-statements this makes no sense because they require constant expressions, but if the expression contains forward-references, it's a hard error anyway and the wasted extra work does no harm.
The pairing of these two parts needs some clean up.