Slight improvement to expr flow
Is it better to fall through on match (faster for first case) vs. pay the branch cost for non-matches (faster for later cases)?
That depends on if we think hex literals will be the most common expr value -- in data, likely, but outside of that constants and labels will be more common in code areas.
Nethertheless, this reads more clearly before and we can test speed better when we get to self-assembly.