Fix a bunch of number parsing
Expr bytecode was writing big-endian but little-endian. Words are written in big-endian order because the expr bytecode and heap are written downwards in RAM, meaning that values appear in little-endian order and would POP in that order too.

Hex parsing was not clearing C for single-digit literals.

printHexByte was not preserving BC, resulting in unwanted sideffects.

There is still some bug where the return value of an expression is incorrect.