Fix handling of labels in expressions
Calculation of the length of an expression when copying to the heap was entirely wrong, additonally A was not 0 as expected because I was using an 8-bit rotation to pop the defer flag instead of an arithmetic shift left.

`deferExpr` already has BC loaded with the expression length, avoiding the need to read expression length from RAM; the goal is to remove the need to store the expression length with the expression entirely.

There is now a bug preventing reading of the second expression of a byte/word keyword!