Fix for symbol search failing
We were using the length of one symbol to compare the next which was failing between symbols with different lengths.
We're supposed to retain the word-length from the word buffer, not re-using the last symbol checked.
The current code is longer than it needs to be and I think a better apporach can be used without the calculated snap-back.