WIP print memory used/free
After assembly we want to print the size of the code-segment (size of generated binary), the size of the heap, the free bytes (between the two) and the peak stack usage.

This reveals that our heap address is being corrupted somewhere (I expect in expression parsing / evaluation).

`osPrintStr` was reworked to not rely on CP/M string termination ("$") and now preseves registers where it didn't before.