Start implementing relative jumps
`stride` is removed and replaced with `param` to hold the parameter length and a flag (bit 7) for relative jumps. If an expression is deferred this byte is now stored in the first reserved byte in the code segment rather than the deferred record which now takes a file number.

A unary negate operator has been added (e.g. -$0010 = $FFF0) but this will need testing to ensure it doesn't conflict with the subtract operator.