Splitting out file functionality to "file.wla"
At the moment the file is automatically opened by the OS bootstrap but we need to be able to support include files and handle pushing/popping the row/col & file-position.
Filenames must be retained (on the heap) as an error in a deferred expression will need to print the file name.