Strings: emit string literals
Strings following a bytes keyword (`.b` ) are emitted to the code-segment as-is.

Some additional work is done to prepare for reading filenames for includes. When reading filenames, the `word` + `expr` buffers are instead used as a forward string buffer for capturing the filename. In the future a separate string buffer will be used as we wish to support used of expressions with include filenames.