Capture include filename string
This reads the filename string but does nothing with it yet.

Progress here will be difficult. We need to keep filename information for error messages but where do we store them? How do we handle the same file being included multiple times?

* RunCPM binary updated to include debugger

This commit made from a Pentium III laptop.