File-reading improvements
Pre-read the first sector of a file.
We no longer need to check the sector read index before reading a character from the sector buffer (DMA @ $80) speeding up reads for 127 consecutive chararcters.

A little extra work is required to return EOF on the NEXT call of getChar where a file ends on a sector boundary, but this logic is only run at each sector read.