Preserve/restore previous file when opening a file
When opening an include file, the previous CP/M FileControlBlock is pushed to the stack so that after the include file is processed, we return to the exact file positon we were in before.

Whilst this is handled for the CP/M FCB, we need to backup the current file number, line & col at the v80 level yet.