Add a dictionary for tracking file-names
Each new file read is assigned a file-number (without duplicates).
If a deferred expression were to error during the 2nd pass, we'd need to be able to return the file-name without having to store the full file-name for each deferred expression.

We still need to restore the current file-name after a file include, but we don't have a way to rewind the file-name record after finding it in the symbol dictionary.