Improve sector reading / restoring
When an FCB is pushed to the stack we now pre-decrement the record number so that the current record can be re-read. This now includes decrementing the extent number when the record number underflows.

The code to read a disk sector after the last character has been read from the current buffer has been improved to be faster in the common case and for the read-sector code to be called from elsewhere.