Show file-name with error messages:
This was way more work than it would seem like. We have to preserve file-names when doing includes and we don't want to push _another_ 12 bytes to the stack on top of the 34+ for the FCB, so we push each file-name to the heap as a symbol so that we can avoid duplication (if a file is included multiple times, like a macro).

There is some annoying back/forth converting between forward strings and backwards string (on-heap) that could so with some refinement. We capture the string forwards; in the future we may allow expressions in strings so we need to mimic code-segment order. The string is reversed to do the symbol search and adding to the heap if necessary and this provides an addr (file ID) that is used to reference the current file-name for errors and deferred expressions.

In order to include the file-name of the first source file, we could no longer rely upon the CP/M pre-built FCB for the 1st parameter. The command-tail is now parsed for the first file-name so that the file is opened the same as any other.

- Error messages show line/col (in hexadecimal ATM)
- Fix bug with ASCII TAB accidentally returning end-of-file
- Fixed error with validating CP/M file-names