Rework file opening/closing
Whilst we can overwrite one CP/M FileControlBlock with another, we should treat it as opaque data.
Instead of overwriting the deafult FCB we now allocate a new FCB on the stack when a file is opened and exclusively use that FCB until the file is closed.
This increases stack usage, particularly as opening the first file allocates an FCB on the stack instead of using the default one in zero-page, but I intend to increase the stack size anyway and there may be other workarounds for this.

Ultimately, allocating new FCBs and not overwriting them does slightly simplify and compact the code.