Make the OS do the command-line parameter parsing
It was a huge pain working out how best to open the initial source file and whether the app or the OS should do it. The CP/M "command tail" is lost after reading a sector so we can't open the source file as soon as we read the parameter but have to hold it until we've read the optional 2nd parameter.

The output file name has to be held until the end and won't use the usual open/close routines as we won't be streaming to it; the write-out will be done in one go by the OS.