Use line/col number from the start of the last word for errors
Instead of the actual line/col of last char read as that can be the end-of-line.

This raises the question of how to report expression errors, particularly deferred ones. Since expressions are parsed into bytecode before evaluation, the error line/col will probably be that of the last word in the expr; not very useful.

We could store the col number with each token, but that may be excessive as some tokens will never produce an error (unary operators).