Report line/col number of error in expr
Because expressions are parsed fully into bytecode before evaluation, errors in expressions were being reported at the line/col of the last word rather than the first.
The line/col number of the start of the expr is now recorded (and pushed in defer records).
Ideally the line/col number of the actual word that caused the error would be preferred, rather than just the start of the expression, but that is somewhat more complicated and the current implementation will suffice for now.