Add a specific error message for using strings with `.w` (words)
Strings are a  list of bytes; can't use a string with the `.w` keyword.

Error messages are taking up a lot of bytes. In the future I may consider putting error messages and full descriptions in a separate file that v80 will read from but that will wait at least until self-assembley is complete first.