Optimisations to evaluation
Some tricky reworking of register selection for the benefit of expression evaluation.
We want to optimise for two things, reading of number literals and label values. Optimising for operators is not as important as there will be half as many operators as there will be operands just by the nature of infix operators.

I believe that there's a bigger gain to be had with the handling of recursive evaluation (parenthesis) by updating the running total by-reference... we'll see.