Final eval optimisations:
Much more performance could be gained out of not using the expression parser for single values, but that will be addressed in the future.

evalExpr and evalValue have been rejigged to use the registers most optimal for fast reading / output of their values. For evalValue this is DE due to the MUL/DIV operators and the proximity of HL, BC instructions such as ADD HL, BC. For evalValue this is BC due to the need to read a label value via a pointer HL->DE->BC.