Echo symbols as they are defined
We now echo a symbol file whilst the source is assembling.

A major oversight in constants was fixed -- getting the value would overwrite the held const name if another label/const appeared in the expression. The const name is pushed to the heap ahead of time and the symbol record completed after obtaining the value.

Various small optimisations to some printing to save chars. That said, the addition of the comment bar and space padding added a lot of bytes.

We have finally broken the 7'000 byte barrier, although I am not pleased about this. My wife said that I'm "fighting over the distance between an eye and an eyelid".

CPM: 753 bytes
ERR: 538 bytes (!)
EXPR: 678 bytes
EMIT: 123 bytes
Z80: 3'715 bytes
CODE: 1'255 bytes