Correct symbol printing
Labels have to be printed value first, then label name because the hex number sets the virtual program-counter, and the label name gets the value.
A by-product of this is some compaction of code, but it highlighted a bug that deferred labels were not being printed. This to be resolved next.

Define label / const swapped positions in the parser to avoid a JR becoming a JP.

`getLabel` returns early for deferred labels, allowing the flag & value to be filled in easily in define label.

Debug label printing removed.