Add relative jump test and fix a bunch of parsing
This adds a relative jump test assembled both by WLA "jr.wla" & v80 "jr.v80" and the binaries are compared.
This highlighted an off-by-one calculation in realtive jumps and a compltely faulty bit of logic when it came to emitting parameter bytes.

Parsing of `.b` keyword was broken in more ways than one. Additional data following a string is correctly supported.