Realign to half-page to save RAM
I couldn't find any bytes to save in the code, but since we've passed the `$xx00` mark in RAM, I've moved the position of the string buffer to `$xx80` to give us 128 more bytes of code whilst saving 128 bytes in RAM as we're no longer aligning to a full page.
Once (if ever?) we pass the `$xx80` boundary with code, we can swap the positions of the `str` and `word`+`expr` buffers again.