Fixed
In a spectacular case of not seeing the trees for the woods, the expression parser wasn't evaluating anything more than a single operator and ignoring the rest of the expression.
Unary operators were also not working due to expected registers changing at some point.

Some code has been rearranged and slightly changed to match my v80 version of v80 (not in repo yet) so that a byte-for-byte comparison can be done when building -- the WLA-DX built binary is padded to 128 bytes to match CP/M for the express reason.