Fix #8: Default to ".v80" if extension not given
This turned out to be absolute hell to do. Every time I had written a solution, an edge case meant I had to scrap it and start it again. It has taken 4 DAYS to fix this "simple" behaviour.

At the moment I have decided not to automatically append a file-extension to an output file-name when present, otherwise there would be no way to output to a binary file without file-extension, and that may be needed. If no output file-name is provided, then the input file-name is used with the extension added or changed to ".COM" (for CP/M)