Rework self-assemble structure
For CP/M compatibility we have to put all files in the same directory and limit names to 8.3.
As a cross-assembler, v80 will support multiple host platforms and target ISAs.
Includes have been redone to start a build from a 'system-ISA' file, in this case "cpm_z80.wla".
Future versions will target 6502 CPUs (e.g. "cpm_6502") and other platform/OSes e.g. "agon_z80", "zeal_z80".
As for native ports, v1 is Z80 code, so all of the above assume running on a Z80 CPU. In the future a separate folder will be used for a native 6502 port, with the same style of sys-ISA includes, such as "c64_6502" (running on C64, assembling 6502 code).