Assemble 6502 code!
This was easier than expected given the drastic reduction in complexity of the humble 6502 over Z80.
v80x65 is just 5KB compared to v80 for Z80 @ 8KB.

The build script saw a lot of work to make v80 assemble itself and then assemble the v80x65 cross-assembler using the bootstrapped version of itself.

Files were renamed in v1 to support "build files", a starting source file that sets the constants that defines the OS & ISA before including v80.

Missing from 6502 support is H/NMOS illegal opcodes. Support for 65C02 will appear in the future too.

There is a nasty bug where if a chain of mutliple includes unfolds and a file is exactly a multiple of 128 bytes, it fails. I haven't been able to isolate / test this in such a way to identify the code that is at fault.