All new instruction parser
This adds a brand-new instruction parser that greatly simplifies the static instruction tables.
The 6502 ISA has been converted, but Z80 (much, much larger) needs to be done.
I did not expect any speed or size improvements to come out of this, I did it only to make doing the eZ80 ISA in the future manageable, however results are much better than expected:
The 6502 conversion saved 93 bytes and ~6M cycles!