More Z80 instructions
This has been kicking my arse for a week or two. I keep bumping into a nasty edge case where the apostrophe trips up the matching where one instruction ends with it, but another continues, for example `ld*hl'` & `ld*hl'a`.
Because we don't know the length of the input word and we need to read-ahead when we encounter an apostrophe, we hit a case where the apostrophe has matched on the first line, but the input word continues and the apostrophe will need to be re-matched on the next line even though we've passed it.
I've tried two entirely different approaches and both run into the same problem.
Right now I don't know if I should just encode the apostrophes into the instruction list despite the bloat just so as to have a working Z80 parser and worry about reducing the size later.
We could read the input word in full before walking the list but we won't have enough registers to do that without a bit of swapping.