Fix Z80 apostrophes
This removes the need to encode (most) apostrophes in the instruction list. Any dot can be replaced with an apostrophe and a trailing apostophe is allowed on any instruction.
This required many, many rewrites due to the parser tripping up on edge cases.
The instruction list has been reduced by removing the superfluous apostrophes and expanded by adding instructions up to [but not including] "S.."