Some fumbling
I'm not certain if this is the right apporach. Expressions are not ideal being handled along with the root / line / body scope due to the way expressions end with any non-expr word. It's probably better to have a separate expr jump-table loop, but this would eat even more memory and we may want to go with a single byte lookup into a table of routines rather than a word for every [in-range] ASCII char.