Port ISAv2 to v0 [WLA]
This replaces the tree used in the WLA-DX version of v80 with the new method used in the v80 port of v80 (v1). I'm surprised that just copying the two functions and fixing the syntax actually worked flawlessly the first time!

I'm doing this so as to modify v0 of v80 to use standard Zilog syntax, then I only need to fix the syntax of v1 and copy the 2 ISA routines and the Z80 table forward.