Begin moving parser to Zilog syntax
This involes supporting spaces in the instruction list, with the parser skipping multiple spaces in the input when it encounters a space in the instruction definition.

A "$" in the instruction definition is expanded to any expression in the input, however the current code halts and doesn't do anything with this. When we capture parameters they will have to be held until after the instruction opcode is confirmed. This will be very messy when it comes to deferred expressions!

BTB, I'm in the process of switching to Linux so this commit includes a mass of whitespace line-ending changes. Secondly, I have to now completely remake the build process from scratch for Linux so it'll be a while before I can get back to this branch :(