Cpature instruction parameters
Since parsing standard Zilog instructions can't know the extact opcode (and prefixes) until the parameters are parsed, we capture them and store them to some temporary variables. Once we identify the instruction, we can take the values from there, doing away with the need to call the parser from within `emitOpcode`.

Deferred expressions complicate matters, a lot. That to be handled in due time.