Problem with returning to instruction parsing
We have a fundamental problem here in that after parsing an instruction parameter expression, we need to reuse the last character read by the expression parser, but we've lost the zero-flag that determines if the char is an EOL/EOF value.
Either we need to store this char state, or more easily recreate it.
Storing it after each character read induces a lot of memory writes for an uncommon read, a function is probably better.
I'm also considering reading a line-at-a-time into a buffer, rather than going through extra layers for each character; some normalisation can happen whilst buffering the line (EOL /r/n to /n and EOF $1A to 0).