Rework `parseHexNumber` to terminate on punctuation
To support instructions like `ld A, [IX+$00]` we want to allow more than just whitespace to terminate a hex number. This change worked out quite nicely, inling the ASCII to integer conversion and doing away with routine calls for each digit.

Now I have to work out why the last character read that we worked so hard to preserve is being eaten when we return to the instruction parser :|