Convert remaining Z80 instruction table to standard syntax
This revealed a major problem we're going to have parsing RST and IM instructions as these have numerical parameters but the opcode used depends upon the parameter. The parameter would have to be written in a very specific way, e.g. "$08" and not "$8" or "8" and we don't have a way to use a literal "$" in instructions yet. It may be necessary to have manual work-arounds in the opcode emitter.

Something else is currently broken running the code, but at least the massive instruction list has been converted.