I have dug this hole and I cannot leave
Wasn't clearing the parameter-count after emitting a parameter value. This fixes the incorrect bytes being emitted but now for some reason the `in A [$00]` instruction is saying it exceeds an 8-bit parameter limit (!?) but not if it's the first instruction in the list (!?).

`parseEpxr` now returns the last-read character and the zero-flag that indicates EOL/EOF which will allow for removing some ugly hacks to reconstruct this value.