Minor changes before we make bigger changes
The opcode flags byte needs reworking to support using parameters for `im`, `rst`, `bit`, `res` & `set` opcodes that use numbers as part of their opcode and not as a parameter byte (although these may also exist).

`param` was renamed `width`( and `emitParam` ro `emitValue`) to better indicate that it represents the number of bytes of values being emitted to the code-segment. I would like to rework file organisation around "heap" and "segment" (or "code"/"data") as there's too much in the "asm" module currently (the internal behaviour of symbol search is not important etc.)