Rework opcode flags handling
Changed the layout of bits in the opcode flags byte to allow for them to be popped off one-by-one instead of using BIT instructions.

Calculating a relative jump is separated out and called based on the opcode flags rather than an flag in the [width] parameter.

For some reason RET instructions are no longer parsing.